webapi/tct-systemsetting-tizen-tests/systemsetting/SystemSettingManager_getProperty_missarg.html

<!DOCTYPE html>
<!--

Copyright (c) 2013 Samsung Electronics Co., Ltd.

Licensed under the Apache License, Version 2.0 (the License);
you may not use this file except in compliance with the License.
You may obtain a copy of the License at

    http://www.apache.org/licenses/LICENSE-2.0

Unless required by applicable law or agreed to in writing, software
distributed under the License is distributed on an "AS IS" BASIS,
WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
See the License for the specific language governing permissions and
limitations under the License.



Authors:
        Andrzej Krolikowski <a.krolikowsk@samsung.com>

-->

<html>
<head>
<title>SystemSettingManager_getProperty_missarg</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
</head>
<body>
<div id="log"></div>
<script type="text/javascript">

//==== TEST: SystemSettingManager_getProperty_missarg
//==== LABEL Check if getProperty throws exception when non-optional argument is missing
//==== SPEC Tizen Web API:System:SystemSetting:SystemSettingManager:getProperty M
//==== SPEC_URL https://developer.tizen.org/help/topic/org.tizen.web.device.apireference/tizen/systemsetting.html
//==== TEST_CRITERIA MMA
var t = async_test(document.title),
    exceptionName = "TypeMismatchError";

t.step(function () {

    assert_throws({name: exceptionName},
        function () {
            tizen.systemsetting.getProperty("HOME_SCREEN");
        }, exceptionName + " should be thrown - missing successCallback.");

    t.done();
});

</script>
</body>
</html>
